requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai
scikit-learn
joblib
numpy
python-dotenv
aiofiles